+++ Paramount.Betterclassifieds.Business/Events/IEventRepository.cs
@@ -11,7 +11,8 @@ namespace Paramount.Betterclassifieds.Business.Events
         EventBooking GetEventBooking(int eventBookingId, bool includeTickets = false);
         IEnumerable<EventTicketReservation> GetEventTicketReservationsForSession(string sessionId);
         IEnumerable<EventTicketReservation> GetEventTicketReservations(int ticketId, bool activeOnly);
-        
+        IEnumerable<EventBookingTicket> GetEventBookingTicketsForEvent(int? eventId);
+
         void CreateEventTicketReservation(EventTicketReservation eventTicketReservation);
         void CreateBooking(EventBooking eventBooking);
 
@@ -19,5 +20,6 @@ namespace Paramount.Betterclassifieds.Business.Events
         void UpdateEventBooking(EventBooking eventBooking);
         void UpdateEventTicket(EventTicket eventTicket);
         void CreateEventTicket(EventTicket ticket);
+        
     }
 }