+++ Paramount.Betterclassifieds.Presentation/ViewModels/Events/EventGuestListViewModel.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace Paramount.Betterclassifieds.Presentation.ViewModels.Events
+{
+    public class EventGuestListViewModel
+    {
+        public string GuestFullName { get; set; }
+        public string GuestEmail { get; set; }
+        public List<EventTicketFieldViewModel> DynamicFields { get; set; }
+    }
+}